+++ build.config.js
@@ -32,6 +32,7 @@ const paths = {
 		assets: {
 			interface: ['src/assets/interface', `release/${platform}/src/assets/interface`],
 			weather: ['src/assets/weather', `release/${platform}/src/assets/weather`],
+			labels: ['src/assets/labels', `release/${platform}/src/assets/labels`],
 			move: ['src/assets/move', `release/${platform}/src/assets/move`],
 			favicons: {
 				ico: ['src/assets/favicon.ico', `release/${platform}/src/assets/favicon.ico`],
@@ -43,7 +44,7 @@ const paths = {
 	extension: {
 		manifest: [`src/manifests/${platform}.json`, `release/${platform}/manifest.json`],
 		scripts: {
-			background: ['src/scripts/services/background.js', `release/${platform}/src/scripts/background.js`],
+			serviceworker: ['src/scripts/services/service-worker.js', `release/${platform}/src/scripts/service-worker.js`],
 			storage: ['src/scripts/services/webext-storage.js', `release/${platform}/src/scripts/webext-storage.js`],
 		},
 	},
@@ -198,7 +199,7 @@ function scripts() {
 	}
 
 	if (PLATFORM_ONLINE) copyFile(...paths.online.serviceworker)
-	if (PLATFORM_EXT) copyFile(...paths.extension.scripts.background)
+	if (PLATFORM_EXT) copyFile(...paths.extension.scripts.serviceworker)
 	if (PLATFORM_EXT) copyFile(...paths.extension.scripts.storage)
 }
 
@@ -206,6 +207,7 @@ function assets() {
 	copyDir(...paths.shared.assets.interface)
 	copyDir(...paths.shared.assets.weather)
 	copyDir(...paths.shared.assets.move)
+	copyDir(...paths.shared.assets.labels)
 	copyFile(...paths.shared.assets.favicons.ico)
 	copyFile(...paths.shared.assets.favicons[128])
 	copyFile(...paths.shared.assets.favicons[512])
@@ -317,7 +319,7 @@ async function updateTranslations() {
 	const translations = []
 
 	for (const lang of langs) {
-		if (lang === 'en') {
+		if (lang === 'en' || lang === '.DS_Store') {
 			continue
 		}